Change how transactions are handled
Some ADO.NET provides (looking at you, MSSQL) require that the current
transaction is assigned to the IDbCommand. This messes up my model of
relying on a transaction in the environment, created by ConnectionProvider.

Instead, move to a different approach. Migration`1 is responsible for
creating its own transactions. Up() and Down() create a transaction if
necessary, and the user-supplied Up() or Down() is run inside this.
Execute() uses the current transaction, and use-created IDbCommands can
get it from a property.

This removes the need for IConnectionProvider and ConnectionProvider, so
they go.

Also make sure that VersionProviderBase runs its things in a transaction.