Move back to the 'ConnectionProvider' model
This is where a ConnectionProvider is used by SimpleMigrator to control
transactions, and has a property of type TConnection to give to the
migrations. This is going to be easier for people to implement for other
connection types, rather than forcing them to add ITransactionProvider to
whatevever their connection type is.

The difference between this and what we had before is that the
ConnectionProvider implementation now returns itself as its 'Connection'
property, and so takes control of the transactions - it makes sure that
IDbCommands have the current transaction associated with them. Basically,
it makes IDbConnection behave as I assumed it did before...