Allow schema creation to be disabled by setting SchemaName to null
In some setups, the user which Simple.Migrations is running as may not
have permission to create the schema. Users in this situation can set
this property to false. It is then up to the user to ensure that the
schema is created with the appropriate permissions before running
SimpleMigrator.